refactor(example): refactoring of directory structures
