--- example/components/screens/index.tsx
@@ -1,1 +1,0 @@
-export * from './MainStackScreen';
