+++ example/index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import {AppContainer} from './AppContainer';
+import {App} from './src/App';
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => AppContainer);
+AppRegistry.registerComponent(appName, () => App);
