+++ example/src/components/HermesAnnounce/index.tsx
