+++ example/src/components/ManagingPeripheralList/Item.tsx
