+++ example/src/components/ManagingPeripheralList/index.tsx
