+++ example/src/components/ScannedPeripheralList/Item.tsx
