+++ example/src/components/ScannedPeripheralList/index.tsx
