+++ example/src/components/index.ts
@@ -3,4 +3,3 @@ export * from './HermesAnnounce';
 export * from './Header';
 export * from './ScannedPeripheralList';
 export * from './ManagingPeripheralList';
-export * from './screens';
