+++ example/src/navigations/index.tsx
@@ -1,0 +1,1 @@
+export * from './MainStack';
