+++ example/src/screens/BleServiceScreens/BatteryServiceScreen.tsx
