+++ example/src/screens/BleServiceScreens/hooks/index.ts
