+++ example/src/screens/BleServiceScreens/hooks/type.ts
