+++ example/src/screens/BleServiceScreens/hooks/useAND_UA_651BLE.tsx
