+++ example/src/screens/BleServiceScreens/hooks/useBatteryService.tsx
